Refactor `configapi` based on latest changes to `configkit`.
